pyproject.toml
[project]
name = "brainrotbench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "elevenlabs>=1.57.0",
  "google-generativeai>=0.8.5",
  "marimo>=0.13.0",
  "openai>=1.75.0",
  "python-dotenv>=1.1.0",
  "requests>=2.32.3",
]

[tool.marimo.runtime]
dotenv = [".env"]
